Refactor CLI to load commands by recursively scanning modules
* Add autoload module to recursively scan for instances of classes at runtime
* Move list and run commands into seperate modules under commands
* Add conftest.py for shared fixtures and functions
