+++ tests/autoload/module_a.py
@@ -1,0 +1,6 @@
+from click import command, echo
+
+
+@command()
+def module_a_command() -> None:
+    echo("Command in module_a")
