+++ tests/autoload/submodule/module_b.py
@@ -1,0 +1,6 @@
+from click import command, echo
+
+
+@command()
+def module_b_command() -> None:
+    echo("Command in module_b")
